move views to separate header
