+++ python/operations.cpp
@@ -26,7 +26,7 @@ data.
 :param x: Variable, DataArray, or Dataset to flatten.
 :param dim: Dimension over which to flatten.
 :raises: If the dimension does not exist, or if x does not contain event lists
-:seealso: :py:class:`scipp.sum`
+:seealso: :py:func:`scipp.sum`
 :return: New variable, data array, or dataset containing the flattened data.
 :rtype: Variable, DataArray, or Dataset)");
 }
